+++ CHANGES.rst
@@ -7,7 +7,9 @@ Changelog
 
 Breaking changes:
 
-- *add item here*
+- Adapt tests to `Products.GenericSetup >= 2.0` thus requiring at least that
+  version.
+  [icemac]
 
 New features: